requirements.txt
pandas
numpy
scikit-learn
psycopg2-binary
joblib
streamlit
python-dotenv
